+++ Carbon/Functions/Test-PowerShellIs64Bit.ps1
@@ -29,9 +29,9 @@ function Test-CPowerShellIs64Bit
 
     if( -not $NoWarn )
     {
-        $msg = 'Carbon''s "Test-CPowerShellIs64Bit" function is OBSOLETE and will be removed in the next major ' +
-               'version of Carbon. Use the new "Test-CPowerShell" function in the new Carbon.Core module instead.'
-        Write-CWarningOnce -Message $msg
+        Write-CRefactoredCommandWarning -CommandName $MyInvocation.MyCommand.Name `
+                                        -ModuleName 'Carbon.Core' `
+                                        -NewCommandName 'Test-CPowerShell'
     }
 
     return ([Environment]::Is64BitProcess)