+++ Carbon/Functions/Uninstall-Certificate.ps1
@@ -109,10 +109,7 @@ function Uninstall-CCertificate
 
         if( -not $NoWarn )
         {
-            $msg = 'Carbon''s "Uninstall-CCertificate" function is OBSOLETE and will be removed in the next major ' +
-                   'version of Carbon. Use the "Uninstall-CCertificate" function in the new "Carbon.Cryptography" ' +
-                   'module.'
-            Write-CWarningOnce -Message $msg
+            Write-CRefactoredCommandWarning -CommandName $MyInvocation.MyCommand.Name -ModuleName 'Carbon.Cryptography'
         }
 
         if( $PSCmdlet.ParameterSetName -like 'ByCertificate*' )