+++ Carbon/Functions/Uninstall-WindowsFeature.ps1
@@ -71,7 +71,9 @@ if( -not (Get-Command -Name 'Get-WindowsFeature*' | Where-Object { $_.ModuleName
         Set-StrictMode -Version 'Latest'
         Use-CallerPreference -Cmdlet $PSCmdlet -Session $ExecutionContext.SessionState
 
-        Write-CWarningOnce -Message ('Uninstall-CWindowsFeature is obsolete and will be removed in a future major version of Carbon.')
+        Write-CObsoleteCommandWarning -CommandName $MyInvocation.MyCommand.Name `
+                                      -NewCommandName 'Uninstall-WindowsFeature' `
+                                      -NewModuleName 'ServerManager'
     
         if( -not (Assert-WindowsFeatureFunctionsSupported) )
         {