+++ Carbon/Functions/Unprotect-String.ps1
@@ -134,9 +134,7 @@ filter Unprotect-CString
     
     if( -not $NoWarn )
     {
-        $msg = 'Carbon''s "Unprotect-CString" function is OBSOLETE and will be removed in the next major version of ' +
-               'Carbon. Use the "Unprotect-CString" function in the new "Carbon.Cryptography" module.'
-        Write-CWarningOnce -Message $msg
+        Write-CRefactoredCommandWarning -CommandName $MyInvocation.MyCommand.Name -ModuleName 'Carbon.Cryptography'
     }
 
     Add-Type -AssemblyName 'System.Security'